Add files via upload
The analysis presented in this Notebook is a testament to the author's proficiency in combining SQL and Python to derive insights from the Worldwideimporters database. With MS SQL Server and Azure Data Studio as the primary database management environment, the author leveraged SQL to extract and manipulate relevant data, which was then visualized using Python. The author's expertise in connecting to the database using SQLAlchemy and Pyodbc modules is commendable.

The analysis itself is a comprehensive study that delves into the customers and their orders, the top-selling products, and the supplier categories. The author also analyzed the top 3 products each month, the sales revenue by customer segment, monthly profits, and sales trends. The author astutely noted the significance of each analysis to businesses, making this Notebook an invaluable resource for business owners.

To visualize the data, the author utilized Plotly, Seaborn, and Matplotlib, while Pandas and Numpy were the primary data manipulation tools. The author's meticulous attention to detail, coupled with their expertise in these tools, has resulted in a professional and amazing Notebook that is a testament to their skills and capabilities.